requirements.txt
aiosqlite==0.21.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.31.0
bcrypt==5.0.0
cffi==2.0.0
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.121.2
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.11
Jinja2==3.1.6
loguru==0.7.3
MarkupSafe==3.0.3
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.10.1
python-multipart==0.0.20
sniffio==1.3.1
sqladmin==0.22.0
SQLAlchemy==2.0.44
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
win32_setctime==1.2.0
WTForms==3.1.2
redis